requirements.txt
aniso8601 #==4.0.1
ase #==3.16.0
dash #==1.16.3
dash-html-components #==1.1.1
dnspython #==1.16.0
Flask #==1.0.4
Flask-API #==1.1
#Flask-Cache #==0.13.1
Flask-Caching
flask-paginate #==0.7.1
Flask-PyMongo #==2.3.0
Flask-RESTful #==0.3.8
gunicorn #==19.8.1
h5py #==2.10.0
itsdangerous #==0.24
Jinja2 #==2.11.2
MarkupSafe ==2.0.1
matplotlib #==3.2.1
monty #==3.0.2
numpy #==1.18.4
pandas #==1.0.4
pymatgen #==2020.4.29
pymongo #==3.10.1
python-ternary #==1.0.7
pytz #==2020.1
ruamel.yaml #==0.16.10
scikit-learn #==0.23.2
seaborn #==0.11.0
six #==1.12.0
Werkzeug ==0.15.5
#crystal-toolkit
#git+git://github.com/materialsproject/crystaltoolkit.git
